getHeight remove parseInt and add parseFloat

+++ MarkupEditor/Resources/markup.js
@@ -19586,7 +19586,7 @@
      editor.style['padding-block-start'] = '0px';
      editor.style['padding-block-end'] = '0px';
      const style = window.getComputedStyle(editor, null);
-     const height = parseInt(style.getPropertyValue('height'));
+     const height = parseFloat(style.getPropertyValue('height'));
      editor.style['padding-block-start'] = paddingBlockStart;
      editor.style['padding-block-end'] = paddingBlockEnd;
      return height;
